+++ pkg/trait/util.go
@@ -19,13 +19,10 @@ package trait
 
 import (
 	"context"
-	"fmt"
 	"strings"
 
 	"github.com/apache/camel-k/pkg/apis/camel/v1alpha1"
 	"github.com/apache/camel-k/pkg/client"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	k8sclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -101,58 +98,4 @@ func VisitKeyValConfigurations(
 			}
 		}
 	}
-}
-
-// GetEnrichedSources returns an enriched version of the sources, with all the external content injected
-func GetEnrichedSources(ctx context.Context, c client.Client, e *Environment, sources []v1alpha1.SourceSpec) ([]v1alpha1.SourceSpec, error) {
-	enriched := make([]v1alpha1.SourceSpec, 0, len(sources))
-
-	for _, s := range sources {
-		content := s.Content
-		if content == "" && s.ContentRef != "" {
-			//
-			// Try to check if the config map is among the one
-			// creates for the deployment
-			//
-			sourceRef := s
-			cm := e.Resources.GetConfigMap(func(m *corev1.ConfigMap) bool {
-				return m.Name == sourceRef.ContentRef
-			})
-
-			//
-			// if not, try to get it from the kubernetes cluster
-			//
-			if cm == nil {
-				key := k8sclient.ObjectKey{
-					Name:      s.ContentRef,
-					Namespace: e.Integration.Namespace,
-				}
-
-				cm = &corev1.ConfigMap{
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "ConfigMap",
-						APIVersion: "v1",
-					},
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      s.ContentRef,
-						Namespace: e.Integration.Namespace,
-					},
-				}
-
-				if err := c.Get(ctx, key, cm); err != nil {
-					return nil, err
-				}
-			}
-
-			if cm == nil {
-				return nil, fmt.Errorf("unable to find a ConfigMap with name: %s in the namespace: %s", s.ContentRef, e.Integration.Namespace)
-			}
-
-			content = cm.Data["content"]
-		}
-		newSource := s
-		newSource.Content = content
-		enriched = append(enriched, newSource)
-	}
-	return enriched, nil
 }
